requirements.txt
streamlit
google-generativeai
qdrant-client
python-dotenv
